# Документацию API (Swagger)/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8090
    description: Generated server url
...
paths:
...
  /request/id:
    get:
      tags:
        - Обращения клиентов
      summary: Get a notification by its id
      operationId: get_3
      parameters:
        - name: id
          in: path
          description: id of notification to be searched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid id supplied
        "404":
          description: notification not found
        "200":
          description: Found the notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDTO'
    put:
      tags:
        - Обращения клиентов
      operationId: edit_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RequestDTO'
    delete:
      tags:
        - Обращения клиентов
      operationId: delete_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RequestDTO'
  /product/{id}:
    get:
      tags:
        - Товары
      operationId: get_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
    put:
      tags:
        - Товары
      operationId: edit_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
    delete:
      tags:
        - Товары
      operationId: delete_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
...
  /items/{id}:
    get:
      tags:
        - Создание заказа на клиента
      operationId: get_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemDto'
    put:
      tags:
        - Создание заказа на клиента
      operationId: edit_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemDto'
    delete:
      tags:
        - Создание заказа на клиента
      operationId: delete_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemDto'
...
  /category/{id}:
    get:
      tags:
        - Категории товаров
      operationId: get_9
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDTO'
    put:
      tags:
        - Категории товаров
      operationId: edit_9
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDTO'
    delete:
      tags:
        - Категории товаров
      operationId: delete_9
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDTO'
...
  /request:
    get:
      tags:
        - Обращения клиентов
      operationId: getAll_3
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestDTO'
    post:
      tags:
        - Обращения клиентов
      operationId: add_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RequestDTO'
  /product:
    get:
      tags:
        - Товары
      operationId: getAll_4
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
    post:
      tags:
        - Товары
      operationId: add_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
...
  /items:
    get:
      tags:
        - Создание заказа на клиента
      operationId: getAll_7
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItemDto'
    post:
      tags:
        - Создание заказа на клиента
      operationId: add_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemDto'
...
  /category:
    get:
      tags:
        - Категории товаров
      operationId: getAll_9
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTO'
    post:
      tags:
        - Категории товаров
      operationId: add_9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDTO'
...
components:
  schemas:
    BoardPreset:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        name:
          type: string
        text:
          type: string
        boardType:
          type: string
          enum:
            - BoardType()
            - BoardType()
            - BoardType()
        boardView:
          type: string
          enum:
            - BoardView()
            - BoardView()
        new:
          type: boolean
    Role:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          maxLength: 255
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        active:
          type: boolean
        new:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        userName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        active:
          type: boolean
        password:
          type: string
        activationCode:
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Role'
        admin:
          type: boolean
        new:
          type: boolean
...
    Contractor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        type:
          type: string
          enum:
            - ContractorType()
            - ContractorType()
            - ContractorType()
            - ContractorType()
        email:
          type: array
          items:
            $ref: '#/components/schemas/Email'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        contractorAddress:
          type: array
          items:
            $ref: '#/components/schemas/ContractorAddress'
        new:
          type: boolean
    ContractorAddress:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        description:
          type: string
        address:
          type: string
        fias:
          type: string
        phone:
          type: string
        contactPersone:
          type: string
        new:
          type: boolean
    Email:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        text:
          type: string
        active:
          type: boolean
        new:
          type: boolean
...
    Phone:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          type: string
        text:
          type: string
        active:
          type: boolean
        new:
          type: boolean
...
    RequestDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        number:
          type: string
        contractor:
          $ref: '#/components/schemas/Contractor'
        contractorAddress:
          $ref: '#/components/schemas/ContractorAddress'
        requestStatus:
          $ref: '#/components/schemas/RequestStatus'
        resolution:
          type: string
          enum:
            - DECLINE
            - DONE
        requestType:
          $ref: '#/components/schemas/RequestType'
    RequestStatus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        name:
          type: string
        requestStatusType:
          type: string
          enum:
            - OPEN
            - IN_PROGRESS
            - DONE
            - CLOSED
        new:
          type: boolean
    RequestType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        name:
          type: string
        requestType:
          $ref: '#/components/schemas/RequestType'
        workFlow:
          $ref: '#/components/schemas/WorkFlow'
        archive:
          type: boolean
        new:
          type: boolean
    WorkFlow:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        boardPreset:
          type: array
          items:
            $ref: '#/components/schemas/BoardPreset'
        new:
          type: boolean
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        name:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        new:
          type: boolean
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cost:
          type: number
          format: double
        title:
          type: string
        category:
          $ref: '#/components/schemas/Category'
...
    OrderItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        qty:
          type: number
          format: double
        plainCartDto:
          $ref: '#/components/schemas/PlainCartDto'
    PlainCartDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...